chore: refractor login ui code
